+++ src/main/java/com/example/pingpong/game/dto/result/GameResult1v1.java
@@ -1,4 +1,4 @@
-package com.example.pingpong.game.model.result;
+package com.example.pingpong.game.dto.result;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
