+++ src/main/java/com/example/pingpong/game/repository/GameResultsRepository.java
@@ -1,6 +1,6 @@
 package com.example.pingpong.game.repository;
 
-import com.example.pingpong.game.model.result.GameResults;
+import com.example.pingpong.game.dto.result.GameResults;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GameResultsRepository extends JpaRepository<GameResults, Long> {
