+++ src/main/java/com/example/pingpong/game/service/GameLogicService.java
@@ -1,0 +1,7 @@
+package com.example.pingpong.game.service;
+
+public class GameLogicService {
+
+
+
+}
