+++ src/main/resources/application.yml
@@ -14,18 +14,13 @@ spring:
         jdbc:
           lob:
             non_contextual_creation: true
-  thymeleaf:
-    prefix: classpath:/templates/  # Thymeleaf 템플릿이 위치한 디렉토리
-    suffix: .html  # 템플릿 파일 확장자
-    cache: false  # 개발 중이라면 캐싱을 비활성화
-    mode: HTML5  # HTML5 모드 사용
-  redis:
-    pool:
-      min-idle: 0
-      max-idle: 8
-      max-active: 8
-    port: 6379
-    host: 127.0.0.1
+    redis:
+      port: 6379
+      host: 127.0.0.1
+      pool:
+        min-idle: 0
+        max-idle: 8
+        max-active: 8
 
 logging:
   level:
